refactor component into smaller component
