+++ app/client-app/src/components/project-type/canvas/index.js
@@ -1,0 +1,2 @@
+export * from "./areas";
+export * from "./sections";
